tests/wpt/web-platform-tests/webxr/webGLCanvasContext_makecompatible_contextlost.https.html

<!DOCTYPE html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script src="resources/webxr_util.js"></script>
  <script src="resources/webxr_test_constants.js"></script>
  <canvas id="webgl-canvas"></canvas>
  <script>

    xr_promise_test(
      "A lost webglCanvasContext should not be able to set xr compatibility",
     (t) => {
      return navigator.xr.test.simulateDeviceConnection(TRACKED_IMMERSIVE_DEVICE)
        .then( (controller) => {
          webglCanvas = document.getElementById('webgl-canvas');
          gl = webglCanvas.getContext('webgl', {xrCompatible: true});
          gl.getExtension('WEBGL_lose_context').loseContext();
          return promise_rejects_dom(t, 'InvalidStateError', gl.makeXRCompatible());
        });
    });

  </script>
</body>
